requirements.txt
flask
dnspython
dkimpy